+++ 08/seq_ex.cpp
@@ -4,17 +4,15 @@ using namespace std;
 
 int main()
 {
-    unsigned int n {0};
+    unsigned long long n {0};
 
     for (int i {0}; i < 50'000'000; ++i)
     {
         n += i;
     }
-
     for (int i {0}; i < 50'000'000; ++i)
     {
         n += i;
     }
-
     cout << n;
 }
